requirements.txt
absl-py==0.13.0
aio-pika==6.4.1
aiormq==3.3.1
astor==0.8.1
beautifulsoup4==4.10.0
bert-dp==1.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cryptography==3.4.8
Cython==0.29.14
DAWG-Python==0.7.2
deeppavlov==0.14.0
docopt==0.6.2
fastapi==0.47.1
filelock==3.0.12
gast==0.2.2
geojson==2.5.0
google-pasta==0.2.0
grpcio==1.40.0
h11==0.9.0
h5py==2.10.0
httptools==0.1.2
idna==2.8
importlib-metadata==4.8.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
multidict==5.1.0
nltk==3.4.5
numpy==1.18.0
opt-einsum==3.3.0
overrides==2.7.0
pamqp==2.3.0
pandas==0.25.3
peewee==3.14.8
plumbum==1.7.1
prometheus-client==0.7.1
protobuf==3.18.0
pycparser==2.20
pydantic==1.3
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
pyOpenSSL==19.1.0
pyowm==3.2.0
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2019.1
requests==2.22.0
rpyc==5.0.1
ruamel.yaml==0.15.100
rusenttokenize==0.0.5
sacremoses==0.0.35
scikit-learn==0.21.2
scipy==1.4.1
six==1.16.0
soupsieve==2.2.1
starlette==0.12.9
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorflow-hub==0.7.0
termcolor==1.1.0
tqdm==4.41.1
typing-extensions==3.10.0.2
urllib3==1.25.11
uvicorn==0.11.7
uvloop==0.16.0
websockets==8.1
Werkzeug==2.0.1
wikipedia==1.4.0
wrapt==1.12.1
xeger==0.3.5
yarl==1.6.3
zipp==3.5.0
